Move the class definition to support.py
